Clean up extension entry point
